Test ConnectorsWrappedToken conversions
